# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title"      : "riboxyz-api",
        "version"    : "1.0.0",
        "description": "A programming interface for ribosome data."
    },
    "paths": {
...
        "/structure/profile": {
            "get": {
                "operationId": "rbxz_bend_api_structure_router_structure_profile",
                "summary": "Structure Profile",
                "parameters": [
                    {
                        "in": "query",
                        "name": "rcsb_id",
                        "schema": {
                            "title": "Rcsb Id",
                            "type": "string"
                        },
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RibosomeStructure"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "STRUCTURE"
                ]
            }
        },
        "/structure/mmcif": {
            "get": {
                "operationId": "rbxz_bend_api_structure_router_structure_mmcif",
                "summary": "Structure Mmcif",
                "parameters": [
                    {
                        "in": "query",
                        "name": "rcsb_id",
                        "schema": {
                            "title": "Rcsb Id",
                            "type": "string"
                        },
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "STRUCTURE"
                ]
            }
        },
        "/structure/ptc": {
            "get": {
                "operationId": "rbxz_bend_api_structure_router_structure_ptc",
                "summary": "Structure Ptc",
                "parameters": [
                    {
                        "in": "query",
                        "name": "rcsb_id",
                        "schema": {
                            "title": "Rcsb Id",
                            "type": "string"
                        },
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/RibosomeStructure"
                                    },
                                    "title": "Response",
                                    "type": "array"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "STRUCTURE"
                ]
            }
        },
        "/structure/ligands": {
            "get": {
                "operationId": "rbxz_bend_api_structure_router_structure_ligands",
                "summary": "Structure Ligands",
                "parameters": [
                    {
                        "in": "query",
                        "name": "rcsb_id",
                        "schema": {
                            "title": "Rcsb Id",
                            "type": "string"
                        },
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/RibosomeStructure"
                                    },
                                    "title": "Response",
                                    "type": "array"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "STRUCTURE"
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "AssemblyInstancesMap": {
                "description": "This basically specifies which assembly an instnace of a polymer or a nonpolymer belongs to. \nCertain PDB structures come with more than a single physical model/assembly packaged in the file,\nhence every chain and many ligands might be present in 2 or more instances. \n\nThe RNA/Protein/Ligand suffices to characterizes all instances, yet, to resolve duplicate chains \nprecisely in space, this information is needed.\n\nassemblies{\nrcsb_id \n    nonpolymer_entity_instances{\n      rcsb_nonpolymer_entity_instance_container_identifiers{\n        entity_id\n        comp_id\n        auth_asym_id\n        rcsb_id\n        auth_seq_id\n      }\n    }\n    polymer_entity_instances{\n       rcsb_polymer_entity_instance_container_identifiers {  \n        entity_id\n        asym_id\n        auth_asym_id\n        entry_id\n        entity_id\n      }\n    }\n  }",
                "properties": {
                    "rcsb_id": {
                        "title": "Rcsb Id",
                        "type": "string"
                    },
                    "nonpolymer_entity_instances": {
                        "anyOf": [
                            {
                                "items": {
                                    "$ref": "#/components/schemas/NonpolymerEntityInstance"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Nonpolymer Entity Instances"
                    },
                    "polymer_entity_instances": {
                        "items": {
                            "$ref": "#/components/schemas/PolymerEntityInstance"
                        },
                        "title": "Polymer Entity Instances",
                        "type": "array"
                    }
                },
                "required": [
                    "rcsb_id",
                    "nonpolymer_entity_instances",
                    "polymer_entity_instances"
                ],
                "title": "AssemblyInstancesMap",
                "type": "object"
            },
            "ChemComp": {
                "properties": {
                    "chem_comp_container_identifiers": {
                        "$ref": "#/components/schemas/ChemCompContainerIdentifiers"
                    }
                },
                "required": [
                    "chem_comp_container_identifiers"
                ],
                "title": "ChemComp",
                "type": "object"
            },
...
            "NonpolymerComp": {
                "properties": {
                    "chemp_comp": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/ChemComp"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "drugbank": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/Drugbank"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "rcsb_chem_comp_target": {
                        "anyOf": [
                            {
                                "items": {
                                    "$ref": "#/components/schemas/RcsbChemCompTarget"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Rcsb Chem Comp Target"
                    }
                },
                "required": [
                    "chemp_comp",
                    "drugbank",
                    "rcsb_chem_comp_target"
                ],
                "title": "NonpolymerComp",
                "type": "object"
            },
            "NonpolymerEntityInstance": {
                "properties": {
                    "rcsb_nonpolymer_entity_instance_container_identifiers": {
                        "$ref": "#/components/schemas/NonpolymerEntityInstanceContainerIdentifiers"
                    }
                },
                "required": [
                    "rcsb_nonpolymer_entity_instance_container_identifiers"
                ],
                "title": "NonpolymerEntityInstance",
                "type": "object"
            },
...
            "NonpolymericLigand": {
                "properties": {
                    "chemicalId": {
                        "title": "Chemicalid",
                        "type": "string"
                    },
                    "chemicalName": {
                        "title": "Chemicalname",
                        "type": "string"
                    },
                    "formula_weight": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Formula Weight"
                    },
                    "pdbx_description": {
                        "title": "Pdbx Description",
                        "type": "string"
                    },
                    "number_of_instances": {
                        "title": "Number Of Instances",
                        "type": "integer"
                    },
                    "nonpolymer_comp": {
                        "$ref": "#/components/schemas/NonpolymerComp"
                    }
                },
                "required": [
                    "chemicalId",
                    "chemicalName",
                    "formula_weight",
                    "pdbx_description",
                    "number_of_instances",
                    "nonpolymer_comp"
                ],
                "title": "NonpolymericLigand",
                "type": "object"
            },
            "Polymer": {
                "properties": {
                    "assembly_id": {
                        "title": "Assembly Id",
                        "type": "integer"
                    },
                    "asym_ids": {
                        "items": {
                            "type": "string"
                        },
                        "title": "Asym Ids",
                        "type": "array"
                    },
                    "auth_asym_id": {
                        "title": "Auth Asym Id",
                        "type": "string"
                    },
                    "parent_rcsb_id": {
                        "title": "Parent Rcsb Id",
                        "type": "string"
                    },
                    "src_organism_names": {
                        "items": {
                            "type": "string"
                        },
                        "title": "Src Organism Names",
                        "type": "array"
                    },
                    "host_organism_names": {
                        "items": {
                            "type": "string"
                        },
                        "title": "Host Organism Names",
                        "type": "array"
                    },
                    "src_organism_ids": {
                        "items": {
                            "type": "integer"
                        },
                        "title": "Src Organism Ids",
                        "type": "array"
                    },
                    "host_organism_ids": {
                        "items": {
                            "type": "integer"
                        },
                        "title": "Host Organism Ids",
                        "type": "array"
                    },
                    "rcsb_pdbx_description": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Rcsb Pdbx Description"
                    },
                    "entity_poly_strand_id": {
                        "title": "Entity Poly Strand Id",
                        "type": "string"
                    },
                    "entity_poly_seq_one_letter_code": {
                        "title": "Entity Poly Seq One Letter Code",
                        "type": "string"
                    },
                    "entity_poly_seq_one_letter_code_can": {
                        "title": "Entity Poly Seq One Letter Code Can",
                        "type": "string"
                    },
                    "entity_poly_seq_length": {
                        "title": "Entity Poly Seq Length",
                        "type": "integer"
                    },
                    "entity_poly_polymer_type": {
                        "title": "Entity Poly Polymer Type",
                        "type": "string"
                    },
                    "entity_poly_entity_type": {
                        "title": "Entity Poly Entity Type",
                        "type": "string"
                    },
                    "nomenclature": {
                        "items": {
                            "$ref": "#/components/schemas/UnionEnum"
                        },
                        "title": "Nomenclature",
                        "type": "array"
                    }
                },
                "required": [
                    "assembly_id",
                    "asym_ids",
                    "auth_asym_id",
                    "parent_rcsb_id",
                    "src_organism_names",
                    "host_organism_names",
                    "src_organism_ids",
                    "host_organism_ids",
                    "rcsb_pdbx_description",
                    "entity_poly_strand_id",
                    "entity_poly_seq_one_letter_code",
                    "entity_poly_seq_one_letter_code_can",
                    "entity_poly_seq_length",
                    "entity_poly_polymer_type",
                    "entity_poly_entity_type",
                    "nomenclature"
                ],
                "title": "Polymer",
                "type": "object"
            },
            "PolymerEntityInstance": {
                "properties": {
                    "rcsb_polymer_entity_instance_container_identifiers": {
                        "$ref": "#/components/schemas/PolymerEntityInstanceContainerIdentifiers"
                    }
                },
                "required": [
                    "rcsb_polymer_entity_instance_container_identifiers"
                ],
                "title": "PolymerEntityInstance",
                "type": "object"
            },
...
            "Protein": {
                "properties": {
                    "assembly_id": {
                        "title": "Assembly Id",
                        "type": "integer"
                    },
                    "asym_ids": {
                        "items": {
                            "type": "string"
                        },
                        "title": "Asym Ids",
                        "type": "array"
                    },
                    "auth_asym_id": {
                        "title": "Auth Asym Id",
                        "type": "string"
                    },
                    "parent_rcsb_id": {
                        "title": "Parent Rcsb Id",
                        "type": "string"
                    },
                    "src_organism_names": {
                        "items": {
                            "type": "string"
                        },
                        "title": "Src Organism Names",
                        "type": "array"
                    },
                    "host_organism_names": {
                        "items": {
                            "type": "string"
                        },
                        "title": "Host Organism Names",
                        "type": "array"
                    },
                    "src_organism_ids": {
                        "items": {
                            "type": "integer"
                        },
                        "title": "Src Organism Ids",
                        "type": "array"
                    },
                    "host_organism_ids": {
                        "items": {
                            "type": "integer"
                        },
                        "title": "Host Organism Ids",
                        "type": "array"
                    },
                    "rcsb_pdbx_description": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Rcsb Pdbx Description"
                    },
                    "entity_poly_strand_id": {
                        "title": "Entity Poly Strand Id",
                        "type": "string"
                    },
                    "entity_poly_seq_one_letter_code": {
                        "title": "Entity Poly Seq One Letter Code",
                        "type": "string"
                    },
                    "entity_poly_seq_one_letter_code_can": {
                        "title": "Entity Poly Seq One Letter Code Can",
                        "type": "string"
                    },
                    "entity_poly_seq_length": {
                        "title": "Entity Poly Seq Length",
                        "type": "integer"
                    },
                    "entity_poly_polymer_type": {
                        "title": "Entity Poly Polymer Type",
                        "type": "string"
                    },
                    "entity_poly_entity_type": {
                        "title": "Entity Poly Entity Type",
                        "type": "string"
                    },
                    "nomenclature": {
                        "items": {
                            "$ref": "#/components/schemas/UnionEnum"
                        },
                        "title": "Nomenclature",
                        "type": "array"
                    },
                    "pfam_accessions": {
                        "items": {
                            "type": "string"
                        },
                        "title": "Pfam Accessions",
                        "type": "array"
                    },
                    "pfam_comments": {
                        "items": {
                            "type": "string"
                        },
                        "title": "Pfam Comments",
                        "type": "array"
                    },
                    "pfam_descriptions": {
                        "items": {
                            "type": "string"
                        },
                        "title": "Pfam Descriptions",
                        "type": "array"
                    },
                    "uniprot_accession": {
                        "items": {
                            "type": "string"
                        },
                        "title": "Uniprot Accession",
                        "type": "array"
                    }
                },
                "required": [
                    "assembly_id",
                    "asym_ids",
                    "auth_asym_id",
                    "parent_rcsb_id",
                    "src_organism_names",
                    "host_organism_names",
                    "src_organism_ids",
                    "host_organism_ids",
                    "rcsb_pdbx_description",
                    "entity_poly_strand_id",
                    "entity_poly_seq_one_letter_code",
                    "entity_poly_seq_one_letter_code_can",
                    "entity_poly_seq_length",
                    "entity_poly_polymer_type",
                    "entity_poly_entity_type",
                    "nomenclature",
                    "pfam_accessions",
                    "pfam_comments",
                    "pfam_descriptions",
                    "uniprot_accession"
                ],
                "title": "Protein",
                "type": "object"
            },
            "RNA": {
                "properties": {
                    "assembly_id": {
                        "title": "Assembly Id",
                        "type": "integer"
                    },
                    "asym_ids": {
                        "items": {
                            "type": "string"
                        },
                        "title": "Asym Ids",
                        "type": "array"
                    },
                    "auth_asym_id": {
                        "title": "Auth Asym Id",
                        "type": "string"
                    },
                    "parent_rcsb_id": {
                        "title": "Parent Rcsb Id",
                        "type": "string"
                    },
                    "src_organism_names": {
                        "items": {
                            "type": "string"
                        },
                        "title": "Src Organism Names",
                        "type": "array"
                    },
                    "host_organism_names": {
                        "items": {
                            "type": "string"
                        },
                        "title": "Host Organism Names",
                        "type": "array"
                    },
                    "src_organism_ids": {
                        "items": {
                            "type": "integer"
                        },
                        "title": "Src Organism Ids",
                        "type": "array"
                    },
                    "host_organism_ids": {
                        "items": {
                            "type": "integer"
                        },
                        "title": "Host Organism Ids",
                        "type": "array"
                    },
                    "rcsb_pdbx_description": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Rcsb Pdbx Description"
                    },
                    "entity_poly_strand_id": {
                        "title": "Entity Poly Strand Id",
                        "type": "string"
                    },
                    "entity_poly_seq_one_letter_code": {
                        "title": "Entity Poly Seq One Letter Code",
                        "type": "string"
                    },
                    "entity_poly_seq_one_letter_code_can": {
                        "title": "Entity Poly Seq One Letter Code Can",
                        "type": "string"
                    },
                    "entity_poly_seq_length": {
                        "title": "Entity Poly Seq Length",
                        "type": "integer"
                    },
                    "entity_poly_polymer_type": {
                        "title": "Entity Poly Polymer Type",
                        "type": "string"
                    },
                    "entity_poly_entity_type": {
                        "title": "Entity Poly Entity Type",
                        "type": "string"
                    },
                    "nomenclature": {
                        "items": {
                            "$ref": "#/components/schemas/UnionEnum"
                        },
                        "title": "Nomenclature",
                        "type": "array"
                    }
                },
                "required": [
                    "assembly_id",
                    "asym_ids",
                    "auth_asym_id",
                    "parent_rcsb_id",
                    "src_organism_names",
                    "host_organism_names",
                    "src_organism_ids",
                    "host_organism_ids",
                    "rcsb_pdbx_description",
                    "entity_poly_strand_id",
                    "entity_poly_seq_one_letter_code",
                    "entity_poly_seq_one_letter_code_can",
                    "entity_poly_seq_length",
                    "entity_poly_polymer_type",
                    "entity_poly_entity_type",
                    "nomenclature"
                ],
                "title": "RNA",
                "type": "object"
            },
...
            "RibosomeStructure": {
                "properties": {
                    "rcsb_id": {
                        "title": "Rcsb Id",
                        "type": "string"
                    },
                    "expMethod": {
                        "title": "Expmethod",
                        "type": "string"
                    },
                    "resolution": {
                        "title": "Resolution",
                        "type": "number"
                    },
                    "pdbx_keywords": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Pdbx Keywords"
                    },
                    "pdbx_keywords_text": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Pdbx Keywords Text"
                    },
                    "rcsb_external_ref_id": {
                        "items": {
                            "type": "string"
                        },
                        "title": "Rcsb External Ref Id",
                        "type": "array"
                    },
                    "rcsb_external_ref_type": {
                        "items": {
                            "type": "string"
                        },
                        "title": "Rcsb External Ref Type",
                        "type": "array"
                    },
                    "rcsb_external_ref_link": {
                        "items": {
                            "type": "string"
                        },
                        "title": "Rcsb External Ref Link",
                        "type": "array"
                    },
                    "citation_year": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Citation Year"
                    },
                    "citation_rcsb_authors": {
                        "anyOf": [
                            {
                                "items": {
                                    "type": "string"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Citation Rcsb Authors"
                    },
                    "citation_title": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Citation Title"
                    },
                    "citation_pdbx_doi": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Citation Pdbx Doi"
                    },
                    "src_organism_ids": {
                        "items": {
                            "type": "integer"
                        },
                        "title": "Src Organism Ids",
                        "type": "array"
                    },
                    "src_organism_names": {
                        "items": {
                            "type": "string"
                        },
                        "title": "Src Organism Names",
                        "type": "array"
                    },
                    "host_organism_ids": {
                        "items": {
                            "type": "integer"
                        },
                        "title": "Host Organism Ids",
                        "type": "array"
                    },
                    "host_organism_names": {
                        "items": {
                            "type": "string"
                        },
                        "title": "Host Organism Names",
                        "type": "array"
                    },
                    "assembly_map": {
                        "items": {
                            "$ref": "#/components/schemas/AssemblyInstancesMap"
                        },
                        "title": "Assembly Map",
                        "type": "array"
                    },
                    "mitochondrial": {
                        "title": "Mitochondrial",
                        "type": "boolean"
                    },
                    "proteins": {
                        "items": {
                            "$ref": "#/components/schemas/Protein"
                        },
                        "title": "Proteins",
                        "type": "array"
                    },
                    "rnas": {
                        "items": {
                            "$ref": "#/components/schemas/RNA"
                        },
                        "title": "Rnas",
                        "type": "array"
                    },
                    "other_polymers": {
                        "items": {
                            "$ref": "#/components/schemas/Polymer"
                        },
                        "title": "Other Polymers",
                        "type": "array"
                    },
                    "nonpolymeric_ligands": {
                        "items": {
                            "$ref": "#/components/schemas/NonpolymericLigand"
                        },
                        "title": "Nonpolymeric Ligands",
                        "type": "array"
                    }
                },
                "required": [
                    "rcsb_id",
                    "expMethod",
                    "resolution",
                    "pdbx_keywords",
                    "pdbx_keywords_text",
                    "rcsb_external_ref_id",
                    "rcsb_external_ref_type",
                    "rcsb_external_ref_link",
                    "citation_year",
                    "citation_rcsb_authors",
                    "citation_title",
                    "citation_pdbx_doi",
                    "src_organism_ids",
                    "src_organism_names",
                    "host_organism_ids",
                    "host_organism_names",
                    "assembly_map",
                    "mitochondrial",
                    "proteins",
                    "rnas",
                    "other_polymers",
                    "nonpolymeric_ligands"
                ],
                "title": "RibosomeStructure",
                "type": "object"
            },
...
        }
    },
    "servers": []
}